[LCR-16] refactor: ErrorCode 어노테이션 수정
